+++ test/test_robot.py
@@ -153,6 +153,7 @@ class TestRobot(EnvironmentSetup):
             robot.SetActiveDOFValues(zeros(robot.GetActiveDOF()))
             assert(manip.FindIKSolution(Tmanip,IkFilterOptions.CheckEnvCollisions) is None)
             assert(manip.FindIKSolution(Tmanip,IkFilterOptions.CheckEnvCollisions|IkFilterOptions.IgnoreEndEffectorCollisions) is not None)
+            assert(not manip.CheckEndEffectorCollision(Tmanip))
 
             box = RaveCreateKinBody(env,'')
             box.InitFromBoxes(array([[0,0,0,0.05,0.05,0.2]]),True)
@@ -165,8 +166,9 @@ class TestRobot(EnvironmentSetup):
             Tmanip = manip.GetTransform()
             robot.SetActiveDOFValues(zeros(robot.GetActiveDOF()))
             assert(manip.FindIKSolution(Tmanip,IkFilterOptions.CheckEnvCollisions|IkFilterOptions.IgnoreEndEffectorCollisions) is not None)
+            assert(not manip.CheckEndEffectorCollision(Tmanip))
 
-            robot.SetActiveDOFValues([ 0.00000000e+00,   9.41227019e-01,   2.95911693e+00, -1.57009246e-16,   0.00000000e+00,  -3.14018492e-16, 0.00000000e+00])
+            robot.SetActiveDOFValues([ 0.00000000e+00,   0.858,   2.95911693e+00, -1.57009246e-16,   0.00000000e+00,  -3.14018492e-16, 0.00000000e+00])
             Tmanip = manip.GetTransform()
             assert(manip.FindIKSolution(Tmanip,IkFilterOptions.CheckEnvCollisions|IkFilterOptions.IgnoreEndEffectorCollisions) is not None)
 
@@ -195,6 +197,12 @@ class TestRobot(EnvironmentSetup):
             box2.SetTransform(manip2.GetTransform())
             robot.Grab(box2,grablink=manip2.GetEndEffector())
             assert(not manip2.CheckEndEffectorCollision(manip2.GetTransform()))
+
+            robot.Grab(target)
+            Tmanip = manip.GetTransform()
+            assert(not manip.CheckEndEffectorCollision(Tmanip))
+            robot.SetActiveDOFValues([ 0.00000000e+00,   0.858,   2.95911693e+00, -1.57009246e-16,   0.00000000e+00,  -3.14018492e-16, 0.00000000e+00])
+            assert(not manip.CheckEndEffectorCollision(Tmanip))
 
 # def test_ikgeneration():
 #     import inversekinematics